https://github.com/redhat-gpe/OPEN_Reporting/issues/276

+++ infrastructure/ansible/roles/utilization_sync/templates/utilization_sync.sh
@@ -6,10 +6,27 @@ echo -en "PGPASSWORD=$PGPASSWORD\n"
 
 echo -en "pull utilization data: $1\n"
 psql -h {{ utilization_db_host }} -U {{ utilization_db_username  }} -w $1 -c "\copy provisioned_services TO '{{database_update_script_dir}}/$1-db.csv' CSV DELIMITER ','"
+if [ $? != 0 ];
+then
+    exit 1;
+fi
 
 echo -en "load utilization data: $1\n"
 mysql -u root {{lms_reporting_db_name}} -e "truncate table lms_reporting.$1;"
+if [ $? != 0 ];
+then
+    exit 1;
+fi
+
 mysql -u root {{lms_reporting_db_name}} -e "load data local infile '{{database_update_script_dir}}/$1-db.csv' into table lms_reporting.$1 fields terminated by ',' lines terminated by '\n';"
+if [ $? != 0 ];
+then
+    exit 1;
+fi
 
 echo -en "run procedure to update dashbuilder"
 mysql -u root {{lms_reporting_db_name}} -e "call refresh_labs_demos;"
+if [ $? != 0 ];
+then
+    exit 1;
+fi
